requirements.txt
flask == 2.0.2
flask-login == 0.5.0
flask-sqlalchemy == 2.5.1
sqlalchemy == 1.4.28
dash == 2.0.0
pandas == 1.3.5
python-dotenv == 0.19.2
plotly == 5.4.0
paho-mqtt == 1.6.1
picamera
requests